refactor: layout의 empty 폴더로 이동
