+++ src/components/layout/empty/EmptyCard/EmptyCard.module.scss
